backend/src/requirements.txt
Django==1.11.4
django-environ==0.4.3
psycopg2==2.7.3
pytz==2017.2
six==1.10.0
djangorestframework==3.7.1
factory-boy==2.9.2
Faker==0.8.4
boto3==1.4.7
boto==2.48.0
botocore==1.7.33
celery==4.1.0
certifi==2017.7.27.1
django-celery-beat==1.0.1
django-cors-headers==2.1.0
django-storages==1.6.5
django-debug-toolbar
django-filter==1.1.0
pycurl==7.43.0
